add the question about card

+++ src/main/java/com/example/present_new/Present.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Present {
-    public List <Integer> price = new ArrayList<>();
+    public List <Double> price = new ArrayList<>();
     public List <String> present = new ArrayList<>();
     public String manufacturer;
     public Integer n;
